refactor(reposicao): split god-component AdminReposicaoPromocoes (774→227)
Quebra a página de Promoções em módulos sob src/components/reposicao/promocoes/.
As 2 useQuery (campanhas/fornecedores) e os useMemo de agrupamento por mês
permanecem no pai; a lógica de upload em lote vira um hook dedicado.

- types.ts: constantes (EMPRESA/FORNECEDOR_DEFAULT/MAX_CONCURRENT/ALL/ESTADOS),
  tipos (UploadItem/Campanha/CampanhaComContagem) + helpers estadoBadgeClass/formatPeriodo.
- badges.tsx: confiancaBadge.
- useUploadPromocoes.ts: hook com FileReader + pool de concorrência (MAX_CONCURRENT)
  + retry + contadores derivados; recebe onProcessed p/ invalidar a lista.
- CampanhasFiltros.tsx: filtros (estado/fornecedor/busca).
- CampanhasTable.tsx: tabela agrupada por mês (colapso/vazio/loading).
- UploadDialog.tsx: modal de upload (presentacional, consome o hook).
- +18 testes (6 arquivos): helpers, badges, filtros, tabela, hook (add/remove/
  reset/contadores) e dialog.

tsc --noEmit limpo; lint 0 problemas.
